deliverables/pipeline-solution/code/google_api_search/requirements.txt
azure-functions
azure-storage-blob>=12.0.0
pandas>=1.0.0
requests>=2.20.0